Extract unitTestSuite model
