+++ packages/insomnia-inso/src/db/prompts.js
@@ -1,10 +1,8 @@
 // @flow
-import type { BaseModel, Environment, UnitTestSuite, Workspace } from './types';
+import type { BaseModel, Environment } from './types';
 import { AutoComplete } from 'enquirer';
 import type { Database } from './index';
-import { mustFindSingle, mustFindSingleOrNone } from './index';
-import flattenDeep from 'lodash.flattendeep';
-import { loadApiSpec } from './models/api-spec';
+import { mustFindSingle } from './index';
 
 export const matchIdIsh = ({ _id }: BaseModel, identifier: string) => _id.startsWith(identifier);
 export const generateIdIsh = ({ _id }: BaseModel, length: number = 10) => _id.substr(0, length);
@@ -28,57 +26,6 @@ export const getDbChoice = (
   value: `${message} - ${idIsh}`,
   hint: config.hint || `${idIsh}`,
 });
-
-export function loadWorkspace(db: Database, identifier: string): ?Workspace {
-  return mustFindSingleOrNone(
-    db.Workspace,
-    s => matchIdIsh(s, identifier) || s.name === identifier,
-  );
-}
-
-export function loadTestSuites(db: Database, identifier: string): Array<UnitTestSuite> {
-  const apiSpec = loadApiSpec(db, identifier);
-  const workspace = loadWorkspace(db, apiSpec?.parentId || identifier);
-
-  // if identifier is for an apiSpec or a workspace, return all suites for that workspace
-  if (workspace) {
-    return db.UnitTestSuite.filter(s => s.parentId === workspace._id);
-  }
-
-  // Identifier is for one specific suite; find it
-  const result = mustFindSingleOrNone(
-    db.UnitTestSuite,
-    s => matchIdIsh(s, identifier) || s.name === identifier,
-  );
-
-  return result ? [result] : [];
-}
-
-export async function promptTestSuites(db: Database, ci: boolean): Promise<Array<UnitTestSuite>> {
-  if (ci) {
-    return [];
-  }
-
-  const choices = db.ApiSpec.map(spec => [
-    getDbChoice(generateIdIsh(spec), spec.fileName),
-    ...db.UnitTestSuite.filter(suite => suite.parentId === spec.parentId).map(suite =>
-      getDbChoice(generateIdIsh(suite), suite.name, { indent: 1 }),
-    ),
-  ]);
-
-  if (!choices.length) {
-    return [];
-  }
-
-  const prompt = new AutoComplete({
-    name: 'testSuite',
-    message: 'Select a document or unit test suite',
-    choices: flattenDeep(choices),
-  });
-
-  const [idIsh] = (await prompt.run()).split(' - ').reverse();
-  return loadTestSuites(db, idIsh);
-}
 
 export function loadEnvironment(
   db: Database,
